Improve fixture creation flexibility and usabilty.
