+++ pyproject.toml
@@ -43,3 +43,4 @@ view = 'walk_blog:main'
 load = 'walk_blog:load'
 browse = 'walk_blog:browse'
 showjson = 'walk_blog:showjson'
+fbuild = 'build_fixtures:main'
