+++ test_data/defaults/users.json
@@ -1,0 +1,1 @@
+["Steve", "Alex", "Fred"]
